Moved to using helper functions for all components
