+++ src/components/CurrentlyPlaying.vue
@@ -19,7 +19,7 @@
 </template>
   
 <script>
-
+import { getBannerUrl, getGamePageUrl, getGames } from "@/scripts/utils";
 export default {
   name: "GameView",
   props: ['gameData'],
@@ -33,29 +33,8 @@ export default {
 
   },
   methods: {
-    getBannerUrl(data) {
-      var banner;
-      if(data.steamId != null && data.steamId != ""){
-        var id = data.steamId;
-        banner = "https://cdn.akamai.steamstatic.com/steam/apps/"+id+"/header.jpg";
-      }else{
-        var title = data.title.toLowerCase().replace(/ /g,"_").replace(/'/g, '');
-        banner = "/game_assets/banners/"+title+"_banner.png";
-      }
-      return banner;
-    },
-    getGamePageUrl(data) {
-      var gamePage;
-      if(data.steamId != null && data.steamId != ""){
-        var id = data.steamId;
-        var title = data.title.replace(/ /g,"_").replace(/'/g, '');
-        gamePage = "https://store.steampowered.com/app/"+id+"/"+title+"/";
-      }else{
-        var title = data.title.toLowerCase().replace(/ /g,"-").replace(/'/g, '');
-        gamePage = "https://store.epicgames.com/en-US/p/"+title;
-      }
-      return gamePage;
-    },
+    getBannerUrl,
+    getGamePageUrl,
   },
 };
